requirements.txt
configparser==6.0.0
mysqlclient==2.1.1
django-rest-framework==0.1.0
django-environ==0.11.2
djangorestframework-simplejwt==5.3.1
Pillow==10.2.0
numpy 
pandas
python-dateutil
six 
tzdata
joblib 
scikit-learn 
scipy 
threadpoolctl
seaborn
plotly_express
django-cors-headers